Changed the folder structure and clean the code
