+++ src/models/ner_model/vocab/key2row
@@ -1,0 +1,1 @@
+€
