+++ src/models/ner_model/vocab/vectors.cfg
@@ -1,0 +1,3 @@
+{
+  "mode":"default"
+}
