+++ src/models/ner_model_2/vocab/vectors.cfg
@@ -1,0 +1,3 @@
+{
+  "mode":"default"
+}
