+++ src/models/ner_model_3/vocab/vectors.cfg
@@ -1,0 +1,3 @@
+{
+  "mode":"default"
+}
